fix permissions on firefox
