+++ src/popup/popup.html
@@ -13,9 +13,9 @@
 <header>
   <header-hero></header-hero>
   <div class="header-buttons">
-    <a role="button" id="mark-as-unread-button" class="outline hidden" href="#">
+    <a role="button" data-new-status="todo" id="mark-as-unread-button" class=" hidden" href="#">
       <img class="icon" src="../../images/icon-todo.png"> save for later</a>
-    <a role="button" id="mark-as-finished-button" class="outline hidden" href="#">
+    <a role="button" data-new-status="done" id="mark-as-finished-button" class=" hidden" href="#">
       <img class="icon" src="../../images/icon-done.png"> mark as finished</a>
     <a role="button" id="settings-button" href="#" class="outline settings" title="settings">
       <img src="../../images/gear.svg" alt="settings">
